requirements.txt
langchain
openai
python-dotenv
requests
beautifulsoup4